Changed CredMan GUI around a bit, now it works as expected
the problem with how we were doing it previously meant that when you
pressed the key Ctrl to trigger the insert commands, cmd would receive
Ctrl + W, Ctrl + a, Ctrl + l and so on. with the alt key, cmd does not
distinguish between Alt + W and W on it's own. I also added a button to
insert the data without using the hotkey. this should close up #2